Move RemoteCacheProviderOptions to remote_provider_traits
