+++ src/rust/engine/remote_provider/remote_provider_traits/src/lib.rs
@@ -55,6 +55,21 @@ pub struct RemoteOptions {
   pub batch_api_size_limit: usize,
 }
 
+#[derive(Clone)]
+pub struct RemoteCacheProviderOptions {
+  // TODO: this is currently framed for the REAPI provider, with some options used by others, would
+  // be good to generalise
+  // TODO: this is structurally very similar to `RemoteOptions`: maybe they should be the same? (see
+  // comment in `choose_provider` too)
+  pub instance_name: Option<String>,
+  pub action_cache_address: String,
+  pub root_ca_certs: Option<Vec<u8>>,
+  pub mtls_data: Option<(Vec<u8>, Vec<u8>)>,
+  pub headers: BTreeMap<String, String>,
+  pub concurrency_limit: usize,
+  pub rpc_timeout: Duration,
+}
+
 #[async_trait]
 pub trait ByteStoreProvider: Sync + Send + 'static {
   /// Store the bytes readable from `file` into the remote store
